+++ js/nls/resources.js
@@ -37,15 +37,16 @@ define({
             tool2:"Change Detection",
             layer: "Layer",
             layerText: "Choose your imagery.",
-            maskImageSelector: "Search for individual images",
-            imageSelectorText: "Choose two different images for comparison.",
+            maskImageSelector: "Choose specific image(s)",
+            imageSelectorText: "Choose two different images to compare.",
             changeText: "Configure change detection.",
             maskText: "Configure mask.",
             zoom: "Zoom in to select images.",
             primary: "Primary Image",
             comparison: "Comparison Image",
-            mode: "Mode",
+            mode: "Visualize changes as a",
             method: "Method",
+            changeMethodText:"Calculate changes in",
             positive: "Positive",
             negative: "Negative",
             threshold: "Threshold",
@@ -66,8 +67,8 @@ define({
             method3: "Soil Adjusted Veg. Index",
             method4: "Water Index",
             method5: "Burn Index",
-            method6: "Less than value",
-            method7: "Greater than value",
+            method6: "Less than my threshold",
+            method7: "Greater than my threshold",
             method8: "Single Band",
             dropDown: "Show images in drop down list.",
             transparency: "Transparency(results)",
@@ -78,6 +79,8 @@ define({
             error4: "Cannot perform action for layer.",
             error5: "Services pre 10.2.1 not supported.",
             error6: "No scenes in current extent.",
+            error7: "Please select two different images.",
+            indexText: "Advanced index options",
             date: "Date",
             areaText: "Area Decrease / Increase",
             areaText2: "Burnt / Post Fire Regrowth Area",
@@ -88,7 +91,12 @@ define({
             extent: "Draw polygon(s) to define extent",
             colorPicker: "Color Picker",
             refresh: "Refresh Button",
-            refreshTooltip: "Refresh query based on current extent."
+            refreshTooltip: "Refresh the image list based on the current extent.",
+            colorpickerText: "Choose a color for your mask",
+            sliderText: "Set your threshold",
+            maskModeText: "Which values should I mask?",
+            positiveSliderText: "Set the minimum increase required between the images before the area is shown in green.",
+            negativeSliderText: "Set the minimum decrease required between the images before the area is shown in magenta."
         },
         editor: {
             title: "Editor",
@@ -118,6 +126,7 @@ define({
             error3: "PixelSize of export is restricted to",
             error4: "for this extent.",
             error5: "Please enter a valid numeric value.",
+            error6: "Your image can't be exported at this time.",
             advance: "Advanced save options",
             modeOption1: "Save to portal",
             modeOption2: "Save to disk",
