+++ js/templates/Mask.html
@@ -1,0 +1,61 @@
+<div id="maskSelector" style="display: none;margin-top: 8px;">
+    <input data-dojo-type="dijit/form/CheckBox" id="imageSelectorMask"/> <label for="imageSelectorMask">${config.i18n.imageMask.maskImageSelector}</label>
+</div>
+<div id="selectorDivMask" style="display:none;">
+    <div id="imageRange" style="margin-top: 10px;margin-bottom: 10px;display: inline-block;"></div>
+    <div id="imageCount" style="display:inline-block;"></div>
+    <div>
+        <div  id="imageSelectContainerMask" style="display:none;float: left;">
+            <button  data-dojo-type="dijit/form/Button" type="button" id="dropDownImageListMask"><img alt="${config.i18n.imageMask.dropDown}" src="images/dropdownlist.png" id="switchDisplayImageMask" class="iconHeight"/></button>
+        </div>
+        <div  style="float:right;display:inline-block;">
+            <button data-dojo-type="dijit/form/Button" type="button" id="refreshImageSliderBtnMask"><img alt="${config.i18n.imageMask.refresh}" src="images/refresh.png" class="iconHeight"/></button>
+        </div>
+        <div id="dropDownOptionMask" style="display: none;">
+            <select  id="imageSelectorDropDownMask" data-dojo-type="dijit/form/Select"></select>
+        </div>
+        <div  id='imageSliderDivMask' style="margin-bottom: 2em;">
+        </div> 
+    </div>
+</div>
+<div id="errorDivMask" style="color: #ee0000;"></div>
+<div id="maskDiv">
+    <hr>
+    <div style="font-weight: bolder;margin-bottom: 10px;">${config.i18n.imageMask.maskText}</div> 
+
+    <div style="display: inline-block;">${config.i18n.imageMask.method}: </div><select autofocus="autofocus"  id="methodMask" data-dojo-type="dijit/form/Select">
+    </select>
+    <br/>
+    <div id="bandInputsMask" style="display: none;margin:1em 0;">
+        <table style="border: 0 solid transparent;"><tr><td><div id="bandName1Mask" style="display: inline-block;"></div></td>
+                <td><select  data-dojo-type="dijit/form/Select" id="band1Mask">
+                    </select>
+                </td></tr>
+            <tr id="bandRowTable"><td><div id="bandName2Mask" style="display: inline-block;"></div></td>
+                <td><select data-dojo-type="dijit/form/Select" id="band2Mask">
+                    </select></td></tr></table>
+    </div>
+    <div id="maskMode" style="margin-bottom: 1em;">
+        <div style="display: inline-block;font-size: 85%;">${config.i18n.imageMask.maskModeText} </div><select  id="maskModeList" data-dojo-type="dijit/form/Select" style="font-size: 85%;margin-left: 5px;">
+            <option value="less">${config.i18n.imageMask.method6}</option>
+            <option value="more">${config.i18n.imageMask.method7}</option>
+        </select>
+    </div>
+    <input id="aoiExtentMask" data-dojo-type="dijit/form/CheckBox" /><label for="aoiExtentMask">${config.i18n.imageMask.extent}</label><div data-dojo-type="dijit/form/DropDownButton"  id="colorButton" style="display: inline-block;margin-left: 0.7em;">
+        <span><img alt="${config.i18n.imageMask.colorPicker}" src="images/ColorPicker.png" class="iconHeight" /></span>
+        <div data-dojo-type="dijit/TooltipDialog" id="colorDialog">
+            <div id="colorPalette" data-dojo-type="dijit/ColorPalette" data-dojo-props="palette:'7x10'"></div>
+        </div>
+    </div>  
+    <div style="display: flex;justify-content: center;margin: 1em 0px;">
+        <button  data-dojo-type="dijit/form/Button" type="button" id="maskApply">${config.i18n.imageMask.apply}</button>
+        <button  data-dojo-type="dijit/form/Button" type="button" id="maskClear">${config.i18n.imageMask.clear}</button>
+    </div>
+    <div id="maskSettingsDiv" style="display: none;margin-bottom: 1.5em;">
+        <div style="margin-top:5px;">${config.i18n.imageMask.sliderText}: <span id="sliderCurrentValue" style="color:#008299;font-weight: bold;"></span></div>
+        <div id="maskSlider"></div>
+    </div>
+    <div id="areaValueContainerMask" style="margin-top: 8px;display:none;">
+        <div style="display: inline-block;">${config.i18n.imageMask.areaText3}:</div>&nbsp;&nbsp;<div id="areaValueMask"  style="font-weight: bolder;color: indianred; display: inline-block;"></div>
+    </div>
+</div>
